requirements.txt
# 签到助手 v2.0 依赖

# HTTP客户端
httpx>=0.25.0

# 配置解析
pyyaml>=6.0

# 日志
loguru>=0.7.0

# 任务调度
apscheduler>=3.10.0
